Add bot.py
